refactor: 페이지 폴더 이름 변경
